- Target compile flags in cmake file. - Fix some warnings.

+++ thyme/src/core/utils.cpp
@@ -15,7 +15,7 @@ std::vector<char> Thyme::readFile(const std::filesystem::path& filePath) {
     }
     const auto size = file.tellg();
     file.seekg(0, std::ios::beg);
-    std::vector<char> buffer(size);
+    std::vector<char> buffer(static_cast<uint64_t>(size));
     file.read(buffer.data(), size);
     file.close();
     return buffer;
